Make deployment system more flexible
Instead of passing the result of the build step to the deploy, give the
source and build info to the deploy function directly. It can just call
the build step itself if it wants to, but it can also behave
differently.

This will be useful with the new build scheduler, because we need to
know whether we're deploying to decide whether to save the resulting
image.