Create helper for access control logic
This will let us remove duplication from the access control logic.
It also adds tests for this logic, and uses (imo) a more readable
implementation of the logic.

We also add tests to verify it works as expected. This is very simple
logic but it is critical for secrurity, thus my motive for fussing with
it.