Simplify CLI logic
Now that we have the constellation absraction, it is clear how we can
reduce code duplication in deriving the engine for a flavor.

Also there the `flavour_opt` type was redundnat, because, any type

```
type t =
  | C1 of t'
  | C2 of t'
  | ...
  | Cn of t'
```

is isomorphic to

```
type c = C1 | C2 | ... | Cn
type t = C * t'
```

And we already have the sum `Pipeline.Flavor.t` for our `c`.