Remove abstraction leak from Cluster
Similary to 9d9bc37d3e97ac471783577fde656937a60edbbb this is a case
where we were needlessly leaking configuration data about the
configuration of a deployer.

In this case, we were using a type like

```
type t =
  | `A1 of string
  | `A2 of string
  | ....
  | `An of string
```

and switching on the constructor to decide which implementation of
`Current_docker` to select. But the only difference between the
`Current_docker` implementation was which `docker_context : string
option` they carried. In other words, we had a roundabout way of
selecting a string.

In this change, we just supply the appropriate string during
configuration time, and instead of needing `n` hard-coded applications
of of the `Current_docker.Make`, we just call it on the provided
`docker_context`.